Split BabySqueel::Nodes into separate files
