+++ lib/baby_squeel/nodes.rb
@@ -1,4 +1,7 @@
-require 'baby_squeel/operators'
+require 'baby_squeel/nodes/node'
+require 'baby_squeel/nodes/attribute'
+require 'baby_squeel/nodes/function'
+require 'baby_squeel/nodes/grouping'
 
 module BabySqueel
   module Nodes
@@ -12,7 +15,7 @@ module BabySqueel
         when Arel::Nodes::Function
           Function.new(arel)
         when Arel::Nodes::Node, Arel::Nodes::SqlLiteral
-          Generic.new(arel)
+          Node.new(arel)
         else
           arel
         end
@@ -28,92 +31,6 @@ module BabySqueel
         else
           node
         end
-      end
-    end
-
-    # This proxy class allows us to quack like any arel object. When a
-    # method missing is hit, we'll instantiate a new proxy object.
-    class Proxy < ActiveSupport::ProxyObject
-      # Resolve constants the normal way
-      def self.const_missing(name)
-        ::Object.const_get(name)
-      end
-
-      attr_reader :_arel
-
-      def initialize(arel)
-        @_arel = Nodes.unwrap(arel)
-      end
-
-      def respond_to?(meth, include_private = false)
-        meth.to_s == '_arel' || _arel.respond_to?(meth, include_private)
-      end
-
-      private
-
-      def method_missing(meth, *args, &block)
-        if _arel.respond_to?(meth)
-          Nodes.wrap _arel.send(meth, *Nodes.unwrap(args), &block)
-        else
-          super
-        end
-      end
-    end
-
-    # This is a generic proxy class that includes all possible modules.
-    # In the future, these proxy classes should be more specific and only
-    # include necessary/applicable modules.
-    class Generic < Proxy
-      extend Operators::ArelAliasing
-      include Operators::Comparison
-      include Operators::Equality
-      include Operators::Generic
-      include Operators::Grouping
-      include Operators::Matching
-    end
-
-    class Attribute < Generic
-      def initialize(parent, name)
-        @parent = parent
-        @name = name
-        super(parent._table[name])
-      end
-
-      def in(rel)
-        if rel.is_a? ::ActiveRecord::Relation
-          ::Arel::Nodes::In.new(self, Arel.sql(rel.to_sql))
-        else
-          super
-        end
-      end
-
-      def _arel
-        if @parent.kind_of? BabySqueel::Association
-          join_expression = @parent._arel
-          table = join_expression.find_alias(@parent)
-          table ? table[@name] : super
-        else
-          super
-        end
-      end
-    end
-
-    # See: https://github.com/rails/arel/pull/381
-    class Function < Generic
-      def initialize(node)
-        super
-        node.extend Arel::OrderPredications
-      end
-    end
-
-    # See: https://github.com/rails/arel/pull/435
-    class Grouping < Generic
-      def initialize(node)
-        super
-        node.extend Arel::AliasPredication
-        node.extend Arel::OrderPredications
-        node.extend Arel::Math
-        node.extend Arel::Expressions
       end
     end
   end
