+++ lib/baby_squeel/nodes/node.rb
@@ -1,0 +1,18 @@
+require 'baby_squeel/operators'
+require 'baby_squeel/nodes/proxy'
+
+module BabySqueel
+  module Nodes
+    # This is a generic proxy class that includes all possible modules.
+    # In the future, these proxy classes should be more specific and only
+    # include necessary/applicable modules.
+    class Node < Proxy
+      extend Operators::ArelAliasing
+      include Operators::Comparison
+      include Operators::Equality
+      include Operators::Generic
+      include Operators::Grouping
+      include Operators::Matching
+    end
+  end
+end
